pyproject.toml
[build-system]
requires = ["setuptools>=67.6.0"]

[project]
name = "Syfala-Config"
version = "0.1.0"
description = ""
authors = [{name = "Rémi Georges", email = "remi.georges@protonmail.com"}]
readme = "README.md"
requires-python = ">=3.7"
keywords = ["config"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

dependencies = [
    "inquirerpy >= 0.3.4",
]
